requirements.txt
whisperx
flask
natsort
pydub
pytorch
torchaudio